fix(wyckoff-io): CIF H-M case-insensitive lookup, POSCAR VASP4 float detection, gen_groups path
- CIF: match _symmetry_space_group_name_H-M case-insensitively (table is
  unique under case-folding) so 'F M -3 M' resolves like 'Fm-3m'.
- POSCAR: detect an absent element-symbols line via f64 (not u64) so a
  non-integer line-6 token like '2.0' is flagged as VASP 4, not mis-read
  as a symbol.
- tools/gen_groups.py: write to wyckoff/data/ (was the stale pre-monorepo
  repo-root data/); regenerated output is byte-identical to committed data.

Adds regression tests for the H-M and POSCAR cases; plugin wasm rebuilt.

+++ wyckoff-io/src/poscar.rs
@@ -37,7 +37,10 @@ pub fn parse(input: &str) -> Result<Record, String> {
     if symbols.is_empty() {
         return Err("POSCAR element-symbols line is empty".into());
     }
-    if symbols[0].parse::<u64>().is_ok() {
+    // A numeric first token (integer OR float, e.g. "4" or "2.0") means the
+    // element-symbols line is absent (VASP 4). f64 catches non-integer counts
+    // that a u64 parse would miss and then mis-read as an element symbol.
+    if symbols[0].parse::<f64>().is_ok() {
         return Err("POSCAR has no element-symbols line (VASP 4 format); insert a VASP 5 \
                     symbols line (e.g. 'Na Cl') before the per-species counts line"
             .into());
